serve/pyproject.toml
[project]
name = "hlab-serve"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.112","uvicorn>=0.30","pydantic>=2.8",
  "numpy>=1.26","opencv-python-headless>=4.9","torch==2.3.*",
  "python-multipart>=0.0.9","pyyaml>=6.0","requests>=2.32"
]
